// src/ocpp/v201/types/report-charging-profiles-request.ts
// Generated by scripts/generate-ocpp.ts. Do not edit.

export type ChargingLimitSourceEnumType = "EMS" | "Other" | "SO" | "CSO";
export type ChargingProfilePurposeEnumType =
  | "ChargingStationExternalConstraints"
  | "ChargingStationMaxProfile"
  | "TxDefaultProfile"
  | "TxProfile";
export type ChargingProfileKindEnumType = "Absolute" | "Recurring" | "Relative";
export type RecurrencyKindEnumType = "Daily" | "Weekly";
export type ChargingRateUnitEnumType = "W" | "A";
export type CostKindEnumType =
  | "CarbonDioxideEmission"
  | "RelativePricePercentage"
  | "RenewableGenerationPercentage";

export interface ReportChargingProfilesRequestV201 {
  customData?: CustomDataType;
  requestId: number;
  chargingLimitSource: ChargingLimitSourceEnumType;
  chargingProfile: [ChargingProfileType, ...ChargingProfileType[]];
  tbc?: boolean;
  evseId: number;
}
export interface CustomDataType {
  vendorId: string;
  [k: string]: unknown;
}
export interface ChargingProfileType {
  customData?: CustomDataType;
  id: number;
  stackLevel: number;
  chargingProfilePurpose: ChargingProfilePurposeEnumType;
  chargingProfileKind: ChargingProfileKindEnumType;
  recurrencyKind?: RecurrencyKindEnumType;
  validFrom?: string;
  validTo?: string;
  chargingSchedule:
    | [ChargingScheduleType]
    | [ChargingScheduleType, ChargingScheduleType]
    | [ChargingScheduleType, ChargingScheduleType, ChargingScheduleType];
  transactionId?: string;
}
export interface ChargingScheduleType {
  customData?: CustomDataType;
  id: number;
  startSchedule?: string;
  duration?: number;
  chargingRateUnit: ChargingRateUnitEnumType;
  chargingSchedulePeriod: [
    ChargingSchedulePeriodType,
    ...ChargingSchedulePeriodType[],
  ];
  minChargingRate?: number;
  salesTariff?: SalesTariffType;
}
export interface ChargingSchedulePeriodType {
  customData?: CustomDataType;
  startPeriod: number;
  limit: number;
  numberPhases?: number;
  phaseToUse?: number;
}
export interface SalesTariffType {
  customData?: CustomDataType;
  id: number;
  salesTariffDescription?: string;
  numEPriceLevels?: number;
  salesTariffEntry: [SalesTariffEntryType, ...SalesTariffEntryType[]];
}
export interface SalesTariffEntryType {
  customData?: CustomDataType;
  relativeTimeInterval: RelativeTimeIntervalType;
  ePriceLevel?: number;
  consumptionCost?:
    | [ConsumptionCostType]
    | [ConsumptionCostType, ConsumptionCostType]
    | [ConsumptionCostType, ConsumptionCostType, ConsumptionCostType];
}
export interface RelativeTimeIntervalType {
  customData?: CustomDataType;
  start: number;
  duration?: number;
}
export interface ConsumptionCostType {
  customData?: CustomDataType;
  startValue: number;
  cost: [CostType] | [CostType, CostType] | [CostType, CostType, CostType];
}
export interface CostType {
  customData?: CustomDataType;
  costKind: CostKindEnumType;
  amount: number;
  amountMultiplier?: number;
}